python/requirements.txt
asgiref==3.7.2
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
dnspython==2.6.1
email-validator==2.1.0.post1
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-WTF==1.2.1
gevent==24.2.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mongoengine==0.27.0
packaging==23.2
pluggy==1.4.0
protobuf==4.25.3
PyJWT==2.8.0
pymongo==4.6.2
pytest==8.0.2
pytest-asyncio==0.23.5.post1
python-dotenv==1.0.1
setuptools==69.1.0
temporalio==1.5.0
types-protobuf==4.24.0.20240129
typing_extensions==4.9.0
Werkzeug==3.0.1
WTForms==3.1.2
zope.event==5.0
zope.interface==6.2
